// src/api/model/apiResponseGuestBookCardResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Forgather API
 * 당신을 위한 순간, 흩어지지 않게. Forgather
 * OpenAPI spec version: 2.0.0
 */
import type { ApiResponseGuestBookCardResponseCode } from './apiResponseGuestBookCardResponseCode';
import type { GuestBookCardResponse } from './guestBookCardResponse';

export interface ApiResponseGuestBookCardResponse {
  code?: ApiResponseGuestBookCardResponseCode;
  message?: string;
  data?: GuestBookCardResponse;
}