// src/api/model/apiResponseGuestBookResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Forgather API
 * 당신을 위한 순간, 흩어지지 않게. Forgather
 * OpenAPI spec version: 2.0.0
 */
import type { ApiResponseGuestBookResponseCode } from './apiResponseGuestBookResponseCode';
import type { GuestBookResponse } from './guestBookResponse';

export interface ApiResponseGuestBookResponse {
  code?: ApiResponseGuestBookResponseCode;
  message?: string;
  data?: GuestBookResponse;
}