// src/api/model/guestBookCardResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Forgather API
 * 당신을 위한 순간, 흩어지지 않게. Forgather
 * OpenAPI spec version: 2.0.0
 */
import type { GuestBookCardPhotoResponse } from './guestBookCardPhotoResponse';

export interface GuestBookCardResponse {
  id?: number;
  /** 방문자 닉네임 */
  nickname?: string;
  /** 메세지 */
  message?: string;
  /** 방명록 카드 생성 시각 */
  createdAt?: string;
  /** 방명록 카드 사진 목록 */
  photos?: GuestBookCardPhotoResponse[];
}